feat: add a script that can unwarp a text file
It is intended to be the inverse of the unix program
`fold`. Thus if you were to run the following:
    ```
    fold -s -w 100 some.txt | unwrap.py > some_unwrapped.txt
    ```
Then the contents of the files `some.txt` and `some_unwrapped.txt`
should be the same.

This code was ported from the perl implementation of the utility
here: https://github.com/gbenison/Line-unwrap. Since the author,
Gregory C Benison did not use a license, I'm assuming it to be public
domain.